# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/admin/categories": {
      "get": {
        "operationId": "AdminCategoriesController_getCategories",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminCategories"
        ]
      },
      "post": {
        "operationId": "AdminCategoriesController_createCategory",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "AdminCategories"
        ]
      }
    },
    "/api/admin/categories/{id}": {
      "get": {
        "operationId": "AdminCategoriesController_getCategoryById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminCategories"
        ]
      },
      "put": {
        "operationId": "AdminCategoriesController_updateCategory",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminCategories"
        ]
      },
      "delete": {
        "operationId": "AdminCategoriesController_deleteCategory",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminCategories"
        ]
      }
    },
    "/api/admin/categories/subcategories": {
      "post": {
        "operationId": "AdminCategoriesController_createSubCategory",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSubCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "AdminCategories"
        ]
      }
    },
    "/api/admin/categories/subcategories/{id}": {
      "put": {
        "operationId": "AdminCategoriesController_updateSubCategory",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSubCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminCategories"
        ]
      },
      "delete": {
        "operationId": "AdminCategoriesController_deleteSubCategory",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminCategories"
        ]
      }
    },
    "/api/admin/categories/product-lists": {
      "post": {
        "operationId": "AdminCategoriesController_createProductList",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductListDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "AdminCategories"
        ]
      }
    },
    "/api/admin/categories/product-lists/{id}": {
      "put": {
        "operationId": "AdminCategoriesController_updateProductList",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductListDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminCategories"
        ]
      },
      "delete": {
        "operationId": "AdminCategoriesController_deleteProductList",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminCategories"
        ]
      }
    },
...
    "/api/admin/content/hero-sections": {
      "get": {
        "operationId": "AdminContentController_getHeroSections",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminContent"
        ]
      },
      "post": {
        "operationId": "AdminContentController_createHeroSection",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateHeroSectionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "AdminContent"
        ]
      }
    },
    "/api/admin/content/hero-sections/{id}": {
      "put": {
        "operationId": "AdminContentController_updateHeroSection",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateHeroSectionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminContent"
        ]
      },
      "delete": {
        "operationId": "AdminContentController_deleteHeroSection",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminContent"
        ]
      }
    },
    "/api/admin/content/vendors": {
      "get": {
        "operationId": "AdminContentController_getVendors",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminContent"
        ]
      },
      "post": {
        "operationId": "AdminContentController_createVendor",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVendorDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "AdminContent"
        ]
      }
    },
    "/api/admin/content/vendors/{id}": {
      "put": {
        "operationId": "AdminContentController_updateVendor",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateVendorDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminContent"
        ]
      },
      "delete": {
        "operationId": "AdminContentController_deleteVendor",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminContent"
        ]
      }
    },
    "/api/admin/content/banners": {
      "get": {
        "operationId": "AdminContentController_getBanners",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminContent"
        ]
      },
      "post": {
        "operationId": "AdminContentController_createBanner",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBannerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "AdminContent"
        ]
      }
    },
    "/api/admin/content/banners/{id}": {
      "put": {
        "operationId": "AdminContentController_updateBanner",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBannerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminContent"
        ]
      },
      "delete": {
        "operationId": "AdminContentController_deleteBanner",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminContent"
        ]
      }
    },
    "/api/admin/content/featured-products": {
      "get": {
        "operationId": "AdminContentController_getFeaturedProductsConfig",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminContent"
        ]
      },
      "put": {
        "operationId": "AdminContentController_updateFeaturedProductsConfig",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFeaturedProductsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminContent"
        ]
      }
    },
...
    "/api/categories/{id}": {
      "get": {
        "operationId": "CategoriesController_getCategoryById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Categories"
        ]
      }
    },
...
  },
  "info": {
    "title": "DOiT E-commerce API",
    "description": "Auto-generated API documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateCategoryDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateCategoryDto": {
        "type": "object",
        "properties": {}
      },
      "CreateSubCategoryDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateSubCategoryDto": {
        "type": "object",
        "properties": {}
      },
      "CreateProductListDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateProductListDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateHeroSectionDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateHeroSectionDto": {
        "type": "object",
        "properties": {}
      },
      "CreateVendorDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateVendorDto": {
        "type": "object",
        "properties": {}
      },
      "CreateBannerDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateBannerDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateFeaturedProductsDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}